requirements.txt
selenium==4.*
pytest==7.*
paramiko==3.*
Appium-Python-Client==3.*
requests==2.*
jsonpath
behave
webdriver-manager==4.*
pillow
jsonschema[format]
kdb-python